Add arg to exerciser "af"
Updated the NuAddFile command to take an extra argument that
identifies the source as coming from a resource fork.  You can't
actually use this to create a forked entry without some behind-
the-scenes renaming, because the exerciser uses the disk filename
as the storage filename, and the "add file" code won't be able to
match the entries up.